Done with playersounds.
The soundObject is still in PlayerView, but the property change has changed. It is now "sound" that stands for the propertyChanged and the newVlue is the sound that is going to play. The PlayerView calls the method getPlayerSound(String playerSound). The string playerSound compares to the different Player events constants and returns the correct sound and the view plays it.